+++ app/controllers/all/opening/projects_controller.rb
@@ -6,11 +6,23 @@ class All::Opening::ProjectsController < ApplicationController
     @date = "#{Date::MONTHNAMES[month.to_i]} #{year}"
   end
 
+  def confirmed_next_month
+    month = Date.today.next_month.month
+    year = Date.today.next_month.year
+    redirect_to action: "confirmed", month: month, year: year
+  end
+
   def revised
     authorize Project, :index?
 
     @projects = Conversion::Project.conversion_date_revised_from(month, year)
     @date = "#{Date::MONTHNAMES[month.to_i]} #{year}"
+  end
+
+  def revised_next_month
+    month = Date.today.next_month.month
+    year = Date.today.next_month.year
+    redirect_to action: "revised", month: month, year: year
   end
 
   def download_csv