+++ spec/features/projects/revised_conversion_date/viewing_projects_with_a_revised_converison_date_spec.rb
@@ -17,7 +17,7 @@ RSpec.feature "Viewing projects with a revised conversion date" do
     create(:date_history, project: project_with_matching_date, previous_date: Date.today.at_beginning_of_month, revised_date: Date.today.at_beginning_of_month + 3.months)
     create(:date_history, project: project_with_matching_date, previous_date: Date.today.at_beginning_of_month + 3.months, revised_date: Date.today.at_beginning_of_month + 4.months)
 
-    visit revised_all_opening_projects_path((Date.today + 3.months).month, Date.today.year)
+    visit "/projects/all/opening/revised/#{(Date.today + 3.months).month}/#{(Date.today + 3.months).year}"
 
     expect(page).to have_content I18n.t("project.revised_conversion_date.title", date: (Date.today + 3.months).to_fs(:govuk_month))
     expect(page).to have_content project_with_matching_date.urn